+++ android/src/main/kotlin/com/smileidentity/flutter/Mapper.kt
@@ -35,12 +35,14 @@ fun convertNonNullMapToNullable(map: Map<String, String>): Map<String?, String?>
     map.mapKeys { it.key }
         .mapValues { it.value }
 
-fun FlutterJobType.toRequest() = when(this) {
+fun FlutterJobType.toRequest() = when (this) {
     FlutterJobType.ENHANCEDKYC -> JobType.EnhancedKyc
+    FlutterJobType.DOCUMENTVERIFICATION -> JobType.DocumentVerification
 }
 
 fun JobType.toResponse() = when (this) {
     JobType.EnhancedKyc -> FlutterJobType.ENHANCEDKYC
+    JobType.DocumentVerification -> FlutterJobType.DOCUMENTVERIFICATION
     else -> TODO("Not yet implemented")
 }
 
@@ -57,19 +59,19 @@ fun PartnerParams.toResponse() = FlutterPartnerParams(
     jobType = jobType?.toResponse(),
     jobId = jobId,
     userId = userId,
-    extras = convertNonNullMapToNullable(extras)
+    extras = convertNonNullMapToNullable(extras),
 )
 
 fun FlutterPartnerParams.toRequest() = PartnerParams(
     jobType = jobType?.toRequest(),
     jobId = jobId,
     userId = userId,
-    extras = convertNullableMapToNonNull(extras)
+    extras = convertNullableMapToNonNull(extras),
 )
 
 fun ConsentInfo.toRequest() = FlutterConsentInfo(
     canAccess = canAccess,
-    consentRequired = consentRequired
+    consentRequired = consentRequired,
 )
 
 fun AuthenticationResponse.toResponse() = FlutterAuthenticationResponse(
@@ -96,7 +98,7 @@ fun FlutterEnhancedKycRequest.toRequest() = EnhancedKycRequest(
         jobType = partnerParams.jobType?.toRequest(),
         jobId = partnerParams.jobId,
         userId = partnerParams.userId,
-        extras = convertNullableMapToNonNull(partnerParams.extras)
+        extras = convertNullableMapToNonNull(partnerParams.extras),
     ),
     sourceSdk = "android (flutter)",
     timestamp = timestamp,
@@ -104,5 +106,5 @@ fun FlutterEnhancedKycRequest.toRequest() = EnhancedKycRequest(
 )
 
 fun EnhancedKycAsyncResponse.toResponse() = FlutterEnhancedKycAsyncResponse(
-    success = success
+    success = success,
 )