+++ ios/Classes/Mapper.swift
@@ -2,20 +2,22 @@ import SmileID
 
 extension FlutterEnhancedKycRequest {
     func toRequest() -> EnhancedKycRequest {
-        EnhancedKycRequest(country: country,
-                           idType: idType,
-                           idNumber: idNumber,
-                           firstName: firstName,
-                           middleName: middleName,
-                           lastName: lastName,
-                           dob: dob,
-                           phoneNumber: phoneNumber,
-                           bankCode: bankCode,
-                           callbackUrl: callbackUrl,
-                           partnerParams: partnerParams.toRequest(),
-                           sourceSdk: "ios (flutter)",
-                           timestamp: timestamp,
-                           signature: signature)
+        EnhancedKycRequest(
+            country: country,
+            idType: idType,
+            idNumber: idNumber,
+            firstName: firstName,
+            middleName: middleName,
+            lastName: lastName,
+            dob: dob,
+            phoneNumber: phoneNumber,
+            bankCode: bankCode,
+            callbackUrl: callbackUrl,
+            partnerParams: partnerParams.toRequest(),
+            sourceSdk: "ios (flutter)",
+            timestamp: timestamp,
+            signature: signature
+        )
     }
 }
 
@@ -33,38 +35,46 @@ extension EnhancedKycAsyncResponse {
 
 extension FlutterPartnerParams {
     func toRequest() -> PartnerParams {
-        PartnerParams(jobId: jobId,
-                      userId: userId,
-                      jobType: jobType!.toRequest())
+        PartnerParams(
+            jobId: jobId,
+            userId: userId,
+            jobType: jobType!.toRequest()
+        )
     }
 }
 
 extension FlutterAuthenticationRequest {
     func toRequest() -> AuthenticationRequest {
         let mappedJobType = jobType.toRequest()
-        return AuthenticationRequest(jobType: mappedJobType,
-                              enrollment: mappedJobType == .smartSelfieEnrollment,
-                              updateEnrolledImage: updateEnrolledImage,
-                              jobId: jobId,
-                              userId: userId)
+        return AuthenticationRequest(
+            jobType: mappedJobType,
+            enrollment: mappedJobType == .smartSelfieEnrollment,
+            updateEnrolledImage: updateEnrolledImage,
+            jobId: jobId,
+            userId: userId
+        )
     }
 }
 
 extension AuthenticationResponse {
     func toResponse() -> FlutterAuthenticationResponse {
-        FlutterAuthenticationResponse(success: success,
-                                      signature: signature,
-                                      timestamp: timestamp,
-                                      partnerParams: partnerParams.toFlutterPartnerParams())
+        FlutterAuthenticationResponse(
+            success: success,
+            signature: signature,
+            timestamp: timestamp,
+            partnerParams: partnerParams.toFlutterPartnerParams()
+        )
     }
 }
 
 extension PartnerParams {
     func toFlutterPartnerParams() -> FlutterPartnerParams {
-        FlutterPartnerParams(jobType: FlutterJobType(rawValue: jobType.rawValue),
-                             jobId: jobId,
-                             userId: userId,
-                             extras: [:])
+        FlutterPartnerParams(
+            jobType: jobType?.toResponse(),
+            jobId: jobId,
+            userId: userId,
+            extras: [:]
+        )
     }
 }
 
@@ -73,7 +83,8 @@ extension FlutterJobType {
         switch (self) {
         case .enhancedKyc:
             return JobType.enhancedKyc
-        default: fatalError("Not yet supported")
+        case .documentVerification:
+            return JobType.documentVerification
         }
     }
 }
@@ -83,6 +94,8 @@ extension JobType {
         switch (self) {
         case .enhancedKyc:
             return FlutterJobType.enhancedKyc
+        case .documentVerification:
+            return FlutterJobType.documentVerification
         default: fatalError("Not yet supported")
         }
     }