+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 10.0.0-beta02
+- Support for Document Verification, exposed as a `SmileIDDocumentVerification` Widget
+- Support for SmartSelfie Authentication, exposed as a `SmileIDSmartSelfieAuthentication` Widget
+
 # 10.0.0-beta01
 - Initial release
 - Support for Enhanced KYC (Async)